fix: halftones masks obeys the transformations
It create a new interface for rendering Transformation for
TaskFilterPixelSW (actually TaskHalftone2 & TaskHalftone3).

For better rendering, it lets the transformations of upper layers
be propagated to lower layers (in other words, pass through this
task), as TaskInterfaceTransformationPass does.
Differently of the latter, though, this new interfacte,
TaskInterfaceTransformationGetAndPass, pass the info to this task
too.

HalfTone2 and HalfTone3 creates masks with shapes that are affected
by the upper transformations, and that is the reason to create this
new interface.
So, the lower layers are better rendered due the propagation of the
upper transformations, as the same time as the mask is correctly
build with that info.